+++ src/main/java/ru/vyarus/dropwizard/guice/module/GuiceyRunner.java
@@ -137,7 +137,7 @@ public class GuiceyRunner {
      * Run lifecycle end.
      */
     public void runFinished() {
-        context.lifecycle().applicationRun();
+        context.bundleStarted();
 
         runTime.stop();
         guiceyTime.stop();